refactor(CLI): Seclude Framework CLI script
